+++ src/pytorch_lightning/loops/fit_loop.py
@@ -219,8 +219,7 @@ class FitLoop(Loop[None]):
         self.trainer._call_strategy_hook("on_train_start")
 
     def on_advance_start(self) -> None:  # type: ignore[override]
-        """Prepares the dataloader for training and calls the hooks ``on_epoch_start`` and
-        ``on_train_epoch_start``"""
+        """Prepares the dataloader for training and calls the hook ``on_train_epoch_start``"""
         model = self.trainer.lightning_module
 
         # reset train dataloader
@@ -245,9 +244,6 @@ class FitLoop(Loop[None]):
         self.epoch_progress.increment_ready()
 
         self.trainer._logger_connector.on_epoch_start()
-
-        self.trainer._call_callback_hooks("on_epoch_start")
-        self.trainer._call_lightning_module_hook("on_epoch_start")
 
         self.trainer._call_callback_hooks("on_train_epoch_start")
         self.trainer._call_lightning_module_hook("on_train_epoch_start")
@@ -298,9 +294,6 @@ class FitLoop(Loop[None]):
         # call train epoch end hooks
         self.trainer._call_callback_hooks("on_train_epoch_end")
         self.trainer._call_lightning_module_hook("on_train_epoch_end")
-
-        self.trainer._call_callback_hooks("on_epoch_end")
-        self.trainer._call_lightning_module_hook("on_epoch_end")
 
         self.trainer._logger_connector.on_epoch_end()