+++ src/pytorch_lightning/CHANGELOG.md
@@ -255,6 +255,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Removed the deprecated way to set the distributed backend via the environment variable `PL_TORCH_DISTRIBUTED_BACKEND`, in favor of setting the `process_group_backend` in the strategy constructor ([#14693](https://github.com/Lightning-AI/lightning/pull/14693))
 
 
+- Removed deprecated callback hooks ([#14834](https://github.com/Lightning-AI/lightning/pull/14834))
+  * `Callback.on_configure_sharded_model` in favor of `Callback.setup`
+  * `Callback.on_before_accelerator_backend_setup` in favor of `Callback.setup`
+  * `Callback.on_batch_start` in favor of `Callback.on_train_batch_start`
+  * `Callback.on_batch_end` in favor of `Callback.on_train_batch_end`
+  * `Callback.on_epoch_start` in favor of `Callback.on_{train,validation,test}_epoch_start`
+  * `Callback.on_epoch_end` in favor of `Callback.on_{train,validation,test}_epoch_end`
+  * `Callback.on_pretrain_routine_{start,end}` in favor of `Callback.on_fit_start`
+
+
 - Removed the deprecated device attributes `Trainer.{devices,gpus,num_gpus,ipus,tpu_cores}` in favor of the accelerator-agnostic `Trainer.num_devices` ([#14829](https://github.com/Lightning-AI/lightning/pull/14829))